+++ src/lib/index.ts
@@ -50,3 +50,4 @@ export * from "./messages";
 export * from "./datalist";
 export * from "./dialog";
 export { confirmation, ConfirmationComponent } from "./confirmation";
+export * from "./qrcode";