+++ src/lib/qrcode/store.ts
@@ -1,0 +1,8 @@
+import { writable } from "svelte/store";
+import type QRScanner from "qr-scanner";
+type QRCode = {
+    scanner?: QRScanner;
+    container?: HTMLDivElement;
+    result?: string | null;
+}
+export const qrcodeStore = writable<QRCode>({});