+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## [1.1.0] - 2024-01-26
+
+### Adições
+
+- Adição do export `qrcode`: nele é possível gerar e ler um QR Code, também é possível fechar o leitor
+- Adição do export `QRCodeReaderComponent`: é o responsável por exibir o leitor do QR Code (`QRCodeReaderComponent` e `qrcode` devem ser usados em conjunto)
+
 ## [1.0.1] - 2024-01-23
 
 ### Correções